Refactor methods names
