+++ src/components/MAPI/clusters/ClusterDetail/ClusterDetailWidgetKubernetesAPI.tsx
@@ -1,5 +1,6 @@
 import URIBlock from 'Cluster/ClusterDetail/URIBlock';
 import { usePermissionsForKeyPairs } from 'MAPI/keypairs/permissions/usePermissionsForKeyPairs';
+import { getClusterK8sAPIUrl } from 'MAPI/utils';
 import { OrganizationsRoutes } from 'model/constants/routes';
 import * as capiv1beta1 from 'model/services/mapi/capiv1beta1';
 import { selectOrganizations } from 'model/stores/organization/selectors';
@@ -34,7 +35,7 @@ const ClusterDetailWidgetKubernetesAPI: React.FC<
   const namespace = selectedOrg?.namespace;
 
   const k8sAPIUrl = cluster
-    ? capiv1beta1.getClusterK8sAPIUrl(cluster)
+    ? getClusterK8sAPIUrl(cluster, provider)
     : undefined;
 
   const gettingStartedPath = useMemo(() => {
